+++ spike/version_rev.py
@@ -10,8 +10,8 @@ except:
     revision = "--not determined--"
 ProgramName = 'SPIKE'
 VersionName = 'Development version'
-version = '0.99.32'
-rev_date = '01-04-2023'
+version = '0.99.33'
+rev_date = '04-07-2023'
 
 def report():
     "prints version name when SPIKE starts"